requirements.txt
Flask>=3.0.3
flask_sqlalchemy>=3.1.1
groq>=0.11.0
pollinations>=1.0.2
python-dotenv>=1.0.1
Requests>=2.32.3
